% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getConfigByName}
\alias{getConfigByName}
\title{Experimental recursive get function
Explores all \code{getOption} \code{.}'s like an object
Process order
root
root a
root a b
root a b c}
\usage{
getConfigByName(name, prefix = "SSQLB:")
}
\description{
Experimental recursive get function
Explores all \code{getOption} \code{.}'s like an object
Process order
root
root a
root a b
root a b c
}
