% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{mainTables}
\alias{mainTables}
\title{Creates a list of \code{\link{tabPanels}}}
\usage{
mainTables(..., id, .adv = FALSE, .tabs = NULL)
}
\arguments{
\item{..., .tabs}{The list of tabs to add}

\item{id}{The id of the \code{\link{tabsetPanel}}}

\item{.adv}{Use the long form (\code{FALSE}) or the short form (\code{TRUE}) for tabs

Short format:
A list of key value pairs
\itemize{
\item key:     The title and identifier for the tabPanel
\item value:   A UITable
}

Long form, more options
A list of title, id, and arguments to pass to tabPanel
\itemize{
\item title:  The title and identifier for the tabPanel
\item id:     The id to use for the tableOutput
\item value:  Optional The value that should be sent when tabsetPanel reports that this tab is selected.
\item icon:   Optional icon to appear on the tab.  This attribute is only valid when using a tabPanel within a navbarPage().
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(ShinySQLBrowser)
mainTables(
  list(CookLog = cook, CookItems = cookitems)
, id = "tabset"
)

mainTables(list(
  list(title = "CookLog", id = "CookLog"),
  list(title = "CookItems", id = "CookItems")
), id = "tabset", .adv = T)
}\if{html}{\out{</div>}}}
}
\description{
Creates a list of \code{\link{tabPanels}}
}
