% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{observer}
\alias{observer}
\title{Observer}
\usage{
observer(session, input, id, func)
}
\arguments{
\item{session}{The Shiny session}

\item{input}{The shiny input}

\item{id}{The shiny value to listen to}

\item{func}{What to do}
}
\description{
Observer
}
