% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{observeSwitch}
\alias{observeSwitch}
\title{Observes the switching of tabs to trigger loading of the tables
Put this in the \code{\link{shinyApp}} \code{server} code}
\usage{
observeSwitch(session, input, container)
}
\arguments{
\item{session}{The Shiny session}

\item{input}{The Shiny input}

\item{container}{A container object of \code{\link{UITables}}}
}
\value{
an observer reference class object
}
\description{
Observes the switching of tabs to trigger loading of the tables
Put this in the \code{\link{shinyApp}} \code{server} code
}
\examples{
ui = bootstrapPage(
  theme = bslib::bs_theme(version = 4), # Optional don't use 5 https://github.com/zekrom-vale/ShinySQLBrowser/issues/2
  includeUITable(container)
)
}
\seealso{
\code{\link{UIContainer}}

\code{\link{shinyApp}}
}
