% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{observeTab}
\alias{observeTab}
\title{Run the scripts of the when it is visable
Server side, Client side, and Set HTML table}
\usage{
observeTab(session, input, tab, ..., .commitout = TRUE, .tabs = NULL)
}
\arguments{
\item{session}{A shiny session}

\item{input}{A Shiny input}

\item{tab}{The tab id}

\item{...}{A list of tabPanels values and UITable objects as key value pairs
Observe the changes of the tabs}

\item{..., .tabs}{A list of tab values as keys and \code{\link{UITable}} objects as values}
}
\description{
Run the scripts of the when it is visable
Server side, Client side, and Set HTML table
}
