% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{procData}
\alias{procData}
\title{Process Data}
\usage{
procData(data, con, .NAisNULL = TRUE)
}
\arguments{
\item{data}{A list of columns returned by JS}

\item{con}{The \code{link{pool}} connection}

\item{.NAisNULL}{Treats \code{NA} as \code{NULL}}
}
\value{
A partial SQL string of \code{id comp val op} like \code{ID == 1 and}
}
\description{
Processes the data list \code{set} statement as a string.
}
