% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{procWhere}
\alias{procWhere}
\title{Process Where}
\usage{
procWhere(where, con, .comp = "==", .op = "and", .NAisNULL = TRUE)
}
\arguments{
\item{con}{The \code{link{pool}} connection}

\item{.comp}{Default comparator \code{equal ==}}

\item{.op}{Default operator \code{and &&} excluded in last statement}

\item{.NAisNULL}{Treats \code{NA} as \code{NULL}}
}
\value{
A partial SQL string of \code{id comp val op} like \code{ID == 1 and}
}
\description{
Process Where
}
