% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pullclass}
\alias{pullclass}
\title{Vectorise \code{class(pull(.data, var))}}
\usage{
pullclass(.data, var)
}
\arguments{
\item{.data}{The table to pull from}

\item{var}{The column to pull}
}
\description{
Vectorise \code{class(pull(.data, var))}
}
