% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sendCreate}
\alias{sendCreate}
\title{Send an SQL Create}
\usage{
sendCreate(data, keys, table, con, .autoinc = TRUE)
}
\arguments{
\item{data}{A list of columns returned by JS}

\item{keys}{The primary keys of the table}

\item{table}{The table to operate on}

\item{con}{The \code{link{pool}} connection}
}
\description{
Send an SQL Create
}
