% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sendRead}
\alias{sendRead}
\title{Send an SQL Remove}
\usage{
sendRead(where, table, con, .comp = "==", .op = "&")
}
\arguments{
\item{where}{A list of operations to do returned by JS}

\item{table}{The table to operate on}

\item{con}{The \code{link{pool}} connection}

\item{.comp}{Default comparator \code{equal ==}}

\item{.op}{Default operator \code{and &&} excluded in last statement}
}
\description{
Send an SQL Remove
}
