% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sendUpdate}
\alias{sendUpdate}
\title{Send an SQL Update}
\usage{
sendUpdate(where, data, table, con, .comp = "==", .op = "and")
}
\arguments{
\item{where}{A list of operations to do returned by JS}

\item{data}{A list of columns returned by JS}

\item{table}{The table to operate on}

\item{con}{The \code{link{pool}} connection}

\item{.comp}{Default comparator \code{equal ==}}

\item{.op}{Default operator \code{and &&} excluded in last statement}
}
\description{
Send an SQL Update
}
