% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tabJs}
\alias{tabJs}
\title{Run the JS start the HTML}
\usage{
tabJs(input, tab, set, .onClickOff = "commit")
}
\arguments{
\item{input}{The shiny input}

\item{tab}{The name of the open tab}

\item{set}{A list of \code{\link{tabPanels}} values and \code{\link{UITable}} objects as key value pairs}
}
\description{
Run the scripts of the tables
Client side, and Set HTML table
}
