% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{tableObserver}
\alias{tableObserver}
\title{Send an SQL Update}
\usage{
tableObserver(session, input, id, uitable, .comp = "==", .op = "&")
}
\arguments{
\item{session}{The Shiny session}

\item{input}{The shiny input}

\item{id}{The shiny value to listen to}

\item{uitable}{The table to operate on}

\item{.comp}{Default comparator \code{equal ==}}

\item{.op}{Default operator \code{and &&} excluded in last statement}
}
\description{
Send an SQL Update
}
