% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tabTable}
\alias{tabTable}
\title{Creates a \code{\link{tabPanel}} with a title and the body of \code{\link{tableOutput}} with an id of id}
\usage{
tabTable(title, id, ..., value = title, icon = NULL)
}
\arguments{
\item{title}{The title of the tab to use}

\item{id}{The id of the \code{\link{tableOutput}}}

\item{...}{UI elements to include within the tab}

\item{value}{The Shiny event value for tabset}

\item{icon}{An icon to use}
}
\description{
Creates a \code{\link{tabPanel}} with a title and the body of \code{\link{tableOutput}} with an id of id
}
