% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{toTable}
\alias{toTable}
\title{Convert a tibble or tbl to a HTML table}
\usage{
toTable(this)
}
\arguments{
\item{this}{A \code{\link{UITable}} object}
}
\value{
An HTML table

Columns have the class of {name}-{col}\if{html}{\out{
<table class="table" id="\{name\}">
 <thead>
   <tr>
     <th class="\{name\}-\{col\}" data-type="\{type\}" id="\{name\}-\{col\}" data-col="\{col\}">
     </th>``
     ...
   </tr>
 </thead>
 <tbody>
   <tr>
     <td class="\{name\}-\{col\}" data-col="\{name\}-\{col\}" data-val="\{value\}">
     </td>
     ...
   </tr>
 </tbody>
</table>
}}
}
\description{
Convert a tibble or tbl to a HTML table
}
