% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{class}
\name{UIContainer-class}
\alias{UIContainer-class}
\title{UIContainer A simple container to hold \code{\link{UITable}} objects}
\description{
UIContainer A simple container to hold \code{\link{UITable}} objects
}
\section{Slots}{

\describe{
\item{\code{data}}{The original data used to generate the \code{\link{UITable}}}

\item{\code{tables}}{A list of \code{\link{UITable}}}
}}

