% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{class}
\name{UITable-class}
\alias{UITable-class}
\title{UITable}
\description{
UITable
}
\section{Slots}{

\describe{
\item{\code{con}}{A \code{\link{pool}} connection}

\item{\code{conname}}{The name of the connection, defaults to the name of the pool argument}

\item{\code{id}}{The HTML table id}

\item{\code{name}}{The tale name to connect to}

\item{\code{types}}{A list of the types of the columns, must name the columns}

\item{\code{opt}}{A list of styles

\code{
   td    = list(class="", collapse=" "),
   tr    = list(class="", collapse=" "),
   th    = list(class="", collapse=" "),
   table = list(class="table"),
   tbody = list(class="table-striped"),
   thead = list(class=""),
   use_this_tbl = c("force", "off", "on")
}}

\item{\code{typemap}}{A list of the HTML input tags to generate per type
If a list is passed instead of character it is used as a reference
list(con=Work, table="people", key="ID", val="Name")
\itemize{
\item con: Optional The connection to use, if not provided use the same connection as the table
\item table: The table to access
\item key: The primary key to use as an identifier of the table (value in option)
\item value: The text to display or pick as the user
\item Use key==value if they are the same
}}

\item{\code{typemap}}{A list of the HTML input tags to generate per type}

\item{\code{input}}{A list of HTML input tags to override \code{typemap}}

\item{\code{js}}{The JS value to access currently the value is \code{UITable.{name}}}

\item{\code{tbl}}{The function to access the tbl, use this to modify the results or filter it}

\item{\code{keys}}{The primary keys in a list separated by \code{|,;:}}

\item{\code{autoinc}}{Does the first primary key use \code{AUTO_INCREMENT}?}
}}

