% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{UITable}
\alias{UITable}
\title{Create a UITable object for Database display and modification}
\usage{
UITable(
  con,
  name,
  id = NULL,
  types = list(),
  typemap = list(),
  opt = list(),
  input = list(),
  js = NULL,
  tbl = dplyr::tbl,
  keys = ""
)
}
\arguments{
\item{con}{A \code{\link{pool}} connection}

\item{name}{The tale name to connect to}

\item{types}{A list of the types of the columns, must name the columns}

\item{typemap}{A list of the HTML input tags to generate per type}

\item{opt}{A list of styles}

\item{input}{A list of HTML input tags to override \code{typemap}}
}
\value{
A new \code{\link{UITable}} object
}
\description{
Create a UITable object for Database display and modification
}
