% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vd}
\alias{vd}
\title{Vectorize \code{[[} Ie. $ or x[\link{i}]}
\usage{
vd(lhs, rhs)
}
\arguments{
\item{lhs}{The list to access}

\item{rhs}{A vector of things to access}
}
\value{
A vector of results
}
\description{
Vectorize \code{[[} Ie. $ or x[\link{i}]
}
\examples{
vd(list(a=1,b=2,c=3), c("b","a","c"))

}
