% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ve}
\alias{ve}
\title{Vectorize \code{[} Ie x\link{i}}
\usage{
ve(lhs, rhs)
}
\arguments{
\item{lhs}{The vector to access}

\item{rhs}{A vector of things to access}
}
\value{
A vector of results
}
\description{
Vectorize \code{[} Ie x\link{i}
}
\examples{
ve(c(1,2,3,4), c(1,3,2))
ve(letters, c(26, 5, 11, 18, 15, 13))
}
