% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{includeUITable}
\alias{includeUITable}
\title{Includes js, css, shinyJS dependencies into the shiny UI
Put this in the \code{\link{shinyApp}} \code{UI} code}
\usage{
includeUITable(
  container,
  jqueryCss =
    "https://cdnjs.cloudflare.com/ajax/libs/jquery-confirm/3.3.4/jquery-confirm.min.css",
  jqueryJS =
    "https://cdnjs.cloudflare.com/ajax/libs/jquery-confirm/3.3.4/jquery-confirm.min.js"
)
}
\arguments{
\item{container}{A container object of UITables}

\item{jqueryCss}{The jquery css to use.  Default: v3.3.4 via cloudflare cdnjs. On \code{NULL} exclude it}

\item{jqueryJs}{The jquery js to use.  Default: v3.3.4 via cloudflare cdnjs. On \code{NULL} exclude it}
}
\value{
A list of UI HTML elements required for ShinySQLBrowser
}
\description{
Includes js, css, shinyJS dependencies into the shiny UI
Put this in the \code{\link{shinyApp}} \code{UI} code
}
