% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pullclass}
\alias{pullclass}
\title{Pull class of columns}
\usage{
pullclass(.data, var)
}
\arguments{
\item{.data}{The table to pull from}

\item{var}{The column to pull}
}
\description{
Used in \code{\link{mutate}} to obtain the class of the column.
Vectorised form of \code{class(pull(.data, var))}
}
